rename: 상태관리 폴더구조 변경
